fix: provide a way for default and out-of-tree matchers
